[Bronze V] Title: 이상한 기호, Time: 0 ms, Memory: 1112 KB, Score: 100 point -BaekjoonHub

+++ 백준/Bronze/15964. 이상한 기호/이상한 기호.c
@@ -1,14 +1,14 @@
 #include <stdio.h>
-int golbeng(int a, int b) {
-	int total;
+long long int golbeng(long long int a, long long int b) {
+	long long int total;
 	total = (a + b) * (a - b);
 	return total;
 }
 int main() {
-	int a, b;
-	scanf("%d %d", &a, &b);
-	int temp;
+	long long int a, b;
+	scanf("%lld %lld", &a, &b);
+	long long int temp;
 	temp = golbeng(a, b);
-	printf("%d", temp);
+	printf("%lld", temp);
 	return 0;
 }
